:sparkles: 멤버 명단 업데이트 서비스 작성

+++ src/main/java/msg/team1/Hi/domain/member/repository/MemberRepository.java
@@ -9,7 +9,7 @@ import java.util.List;
 import java.util.Optional;
 
 @Repository
-public interface MemberRepository extends JpaRepository<Member, Integer> {
+public interface MemberRepository extends JpaRepository<Member, Long> {
     Optional<Member> findByEmail(String email);
     Optional<Member> findByName(String name);
     List<Member> findAllByReservation(Reservation reservation);
